implement modules for actions using SOLID best practices
